refactor advisor cockpit acknowledgement route
