+++ src/app/router_registry.py
@@ -1,6 +1,9 @@
 from fastapi import FastAPI
 
 from app.routers.advisor_cockpit import router as advisor_cockpit_router
+from app.routers.advisor_cockpit_acknowledgements import (
+    router as advisor_cockpit_acknowledgements_router,
+)
 from app.routers.advisor_cockpit_house_view import (
     router as advisor_cockpit_house_view_router,
 )
@@ -209,6 +212,7 @@ from app.routers.workbench_sandbox import router as workbench_sandbox_router
 
 def register_routers(app: FastAPI) -> None:
     app.include_router(advisor_cockpit_router)
+    app.include_router(advisor_cockpit_acknowledgements_router)
     app.include_router(advisor_cockpit_projections_router)
     app.include_router(advisor_cockpit_house_view_router)
     app.include_router(bank_demo_proof_router)
